requirements.txt
asgiref==3.3.4
cairocffi==1.0.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
cssselect2==0.4.1
DAWG-Python==0.7.2
defusedxml==0.7.1
Django==3.2
django-admin-rangefilter==0.8.1
django-admin-sortable2==1.0
django-appconf==1.0.4
django-compressor==2.4.1
django-haystack==3.0
django-material-admin==1.8.4
django-mysql==3.11.1
django-recaptcha3==0.4.0
django-summernote==0.8.11.6
django-taggit==1.4.0
django-yaturbo==1.0.0
docopt==0.6.2
html5lib==1.1
idna==2.10
importlib-metadata==4.0.1
josepy==1.8.0
mysqlclient==2.0.3
phonenumbers==8.12.21
Pillow==8.2.0
pkg-resources==0.0.0
pycparser==2.20
pymemcache==3.4.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyOpenSSL==20.0.1
Pyphen==0.10.0
pytils==0.3
pytz==2021.1
razdel==0.5.0
rcssmin==1.0.6
requests==2.25.1
rjsmin==1.1.0
rutermextract==0.3
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
tinycss2==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
WeasyPrint==52.5
webencodings==0.5.1
Whoosh==2.7.4
zipp==3.4.1
